+++ src/app/core/models/event.ts
@@ -31,8 +31,6 @@ export interface IEvent {
   entitiesArray?: string[];
   auditItems?: IBase[];
   userId?: string;
-  usersFavs?: string[];
-  nClaps?: number;
   extra?: string; // Extra field to pass info
 }
 
@@ -70,9 +68,6 @@ export class Event implements IEvent, IBase { // IAudit
     public auditItems?: IBase[],
     public userId?: string,
 
-    public usersFav?: string[],
-    public nClaps?: number,
-
     public extra?: string,
      ) {
   }
@@ -94,7 +89,6 @@ export class Event implements IEvent, IBase { // IAudit
       [], [], // Entity
       [],  // Audit
       null, // UserId
-      [], 0, // userFavs, nClaps
     );
   }